extract snapshots test
