+++ packages/insomnia-cli/package.json
@@ -8,11 +8,11 @@
   "scripts": {
     "typecheck": "",
     "test": "jest",
-    "test:snapshots": "jest -u",
-    "start": "npm run watch",
-    "clean": "rimraf dist",
-    "build": "babel ./src --out-dir=./dist --ignore '**/*/__tests__/**'",
-    "watch": "babel --source-maps --watch=./src --out-dir=./dist --ignore '**/*/__tests__/**'",
+    "test:watch": "jest --watch",
+    "test:snapshots": "npm run build && jest -u",
+    "start": "npm run build:watch",
+    "build": "rimraf dist && babel ./src --out-dir=./dist --ignore '**/*/__tests__/**'",
+    "build:watch": "rimraf dist && babel --source-maps --watch=./src --out-dir=./dist --ignore '**/*/__tests__/**'",
     "bootstrap": "npm run build",
     "prepublish": "npm run build"
   },
